passing feature

+++ discordBot/bot.py
@@ -3,6 +3,7 @@ from dotenv import load_dotenv
 import discord
 from test_utils.testable_bot import TestableBot
 import requests
+import json
 
 # backend currently must be hosted locally
 baseurl = 'https://321-hosted-backend.jack-klob.repl.co'
@@ -38,14 +39,19 @@ async def task_creation(ctx, *args):
 @bot.command(name = 'delete_task', help = 'Deleted a task')
 async def task_delete(ctx, id):
     url = f'{baseurl}/task/{id}'
+    guild_id = ctx.guild.id
+
+    task = json.loads(requests.get(url=f'{baseurl}/task/{id}').text)
+
+    if 'id' not in task or task['guild'] != guild_id:
+        await ctx.send(f'Task with id {id} does not exist')
+        return
+    
     r = requests.delete(url=url)
     print(r)
-    print(r.text)
-
+    
     if r.status_code == 204:
         await ctx.send(f'Task with id {id} deleted')
-    elif r.status_code == 404:
-        await ctx.send(f'Task with id {id} does not exist')
     else:
         await ctx.send("An error occured when trying to delete task")
